Hid admin links and actions from the buyer - also hid buyer type actions from admin

+++ app/actions.rb
@@ -155,15 +155,19 @@ put '/sellers/:id/edit' do
 end
 
 delete '/sellers/:id' do
-  seller = Seller.find params[:id]
-  seller.destroy
-  redirect to '/sellers'
+  unless current_buyer
+    seller = Seller.find params[:id]
+    seller.destroy
+  end
+    redirect to '/sellers'
 end
 
 delete '/products/:id' do
-  product = Product.find params[:id]
-  product.destroy
-  redirect to '/products'
+  unless current_buyer
+    product = Product.find params[:id]
+    product.destroy
+  end
+    redirect to '/products'
 end
 
 post '/sellers/show' do
